requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
